+++ lua/vv-scrollbar/types/state.lua
@@ -3,7 +3,7 @@
 ---@field did_setup boolean
 ---@field bars table<integer, VVScrollbar.Bar>
 ---@field git_marks table<integer, vv-utils.git.DiffLineSets>
----@field git_pending table<integer, boolean>
+---@field git_pending table<integer, { request: vv-utils.async.Request, signature: string, dirty: boolean, schedule_refresh: fun() }>
 ---@field dragging? VVScrollbar.DragState
 ---@field augroup? integer
 ---@field refresh_throttled? fun()